refactor: move kafka-consumers to api module; remove kafka deps in routing module
