+++ ors-api/src/main/java/org/heigit/ors/api/kafka/ORSKafkaConsumer.java
@@ -18,7 +18,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.heigit.ors.kafka;
+package org.heigit.ors.api.kafka;
 
 import java.util.ArrayList;
 import java.util.List;
