+++ ors-api/src/main/java/org/heigit/ors/api/kafka/ORSKafkaConsumerRunner.java
@@ -18,7 +18,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.heigit.ors.kafka;
+package org.heigit.ors.api.kafka;
 
 import org.apache.kafka.clients.consumer.*;
 import org.apache.kafka.common.serialization.LongDeserializer;
