+++ ors-api/src/test/java/org/heigit/ors/apitests/status/ResultTest.java
@@ -16,7 +16,7 @@ package org.heigit.ors.apitests.status;
 import org.heigit.ors.apitests.common.ServiceTest;
 import org.heigit.ors.apitests.common.VersionAnnotation;
 import org.heigit.ors.apitests.common.EndPointAnnotation;
-import org.heigit.ors.kafka.ORSKafkaConsumerRunner;
+import org.heigit.ors.api.kafka.ORSKafkaConsumerRunner;
 import org.junit.jupiter.api.Test;
 import org.opentest4j.AssertionFailedError;
